[SYCL][UR][L0 v2] fix urQueueCreateWithNativeHandle
getSignalEvent in ur_command_list_manager returned nullptr
when queue was not set. This was the case when initializing
the command list manager in queue ctor called from
urQueueCreateWithNativeHandle.

Fix this by passing queue to the command list manager and
removing the default nullptr value from the ctor (to avoid
this issue in future).

Also, event if queue is nullptr, return the event instead
of nullptr from getSignalEvent. This event is mostly functional,
it just cannot be used for profiling.