[SYCL] Fix vec constructors (#17712)
Implements https://github.com/KhronosGroup/SYCL-Docs/pull/668.

Note that the "test" changes here are less of the documentation of
"intent" but rather the documentation of the effect of the change.

Also, the idea behind specification changes was to keep all "reasonable"
code in the wild compiling and several changes are needed together to
make that happen. Until all of them land, keep the new code out of
`-fpreview-breaking-changes` mode.