[Driver][NFC][FPGA] Remove FPGA archive support
Remove the specific FPGA support in the driver and tools that are
associated with FPGA archives.  These archives provided an intermediate
representation of binaries used for FPGA targets.

Creation of these archives were accomplished by using the
-fsycl-link=early and -fsycl-link=image options along with the
-fintelfpga enabling option.

This is simply a step towards full removal of FPGA support in the
compiler driver and clang-offload-bundler.