[SYCL][Graph] Fix issue when updating HostTasks with whole graph update (#17645)
When updating a HostTask using whole graph update, we rely
on the Sycl Scheduler to execute the HostTask and all its
dependencies. The Scheduler executes this work asynchronously
using a separate thread. Since exec_graph.update() was non-blocking,
this was causing issues, where the user code could, for example,
free resources, before the update() command (which is a dependent on the
HostTask completion) could execute.

This commit fixes this issue by making exec_graph.update() blocking
when host_tasks are used.

Co-authored-by: Ewan Crawford <ewan@codeplay.com>