[SYCL][E2E] Fix flaky failure of `Basic/image/image_max_size.cpp` (#17834)
This test requires a significant amount of host memory. It has been
observed that sometimes (very rarely) the test may fail with
OUT_OF_HOST_MEMORY error, especially when run in parallel with other
"high-overhead" tests. Refer CMPLRLLVM-66341.

This PR makes the test ignore failure if that happens. An alternative is
to check for the available host memory and skip the test if it is too
low. However, that approach is still susceptible to race conditions.