[UR] Added "roundtrip" GetInfo comformance tests (#17738)
This tests that converting a UR handle to a native handle, then back
into a UR handle preserves/recreates any handles contained within it.

The event test is a special case; since urEventCreateWithNativeHandle
doesn't accept a queue paramater. For this, we tollerate a different
queue being created, as long as they both have the same underlying
native handle.