[AsyncAlloc][CUDA] Change memory pool max size error to a warning (#17863)
Warn users over setting the memory pool maximum size property instead of
error.

This should aid CI which is affected by the memory_pool.cpp e2e test
which expects to be able to set this property despite DPC++ being built
with a CUDA version that is too old.

See this CI error:
https://github.com/intel/llvm/pull/17757#issuecomment-2778601572

This PR relies on https://github.com/intel/llvm/pull/17095 to be merged
for the printing of the warning to be enabled.