[CI][Benchmarks] update llama.cpp and requirements to latest
This patch updates llama.cpp to the latest available version, uses a
new, more relevant, GGUF model, and updates oneAPI to 2025.1.

I was trying to avoid updating oneAPI, but the latest llama.cpp
internal pooling logic seems to be broken on 2025.0, resulting in
double-free errors when using older oneAPI components.

The utils.download function also had to be updated, because it was
using a deprecated features and didn't work on some configurations.