[AsyncAlloc][SYCL] Introduce trim_to to memory_pool to trim memory allocated to a memory pool
Introduce the memory_pool member function trim_to which explicitly trims
allocated memory for a memory_pool. It takes a single parameter which
specifies the number of bytes to keep after the trimming operation.

The implementation may or may not trim down to the specified number of
bytes, and will not trim memory that is allocated to ptrs for the user.