[LIT][E2E] Fix LIT hang when executing non-existent files (#17505)
Fixes #16351

In our containers if an exception is not immediately caught lit will
hang on the following call
https://github.com/intel/llvm/blob/4adaef0e444cf9a7e11f433b9995ece6bf9e0aa4/llvm/utils/lit/lit/run.py#L93

This can occur when using the internal lit shell and trying to run a
program that does not exist. In this case `_executeShCmd` will throw an
internal shell error, which will not be caught by the function directly
calling it, `executeShCmd`, rather it is caught one function higher in
the call stack in `executeScriptInternal`. Because that exception is
percolated up the call stack instead of being immediately caught lit
will hang. This patch changes the location where we catch this exception
to `executeShCmd` instead to avoid this.

Previously to avoid this we would use the external lit shell. However
this introduces some differences in how we need to write tests (i.e.,
needing to add `--crash` for certain `not` calls), it slightly changes
how test output is printed (all in one block, rather than separated by
`RUN:` lines), and it messes up the path to executables when running on
Windows (all `\` were interpreted as escapes for the next characters
leading to trying to execute a non-existent file). This pr also changes
the E2E tests to always use the internal lit shell.

For more background on what causes this hang see:

https://stackoverflow.com/questions/15314189/python-multiprocessing-pool-hangs-at-join
https://bugs.python.org/issue9400
https://github.com/python/cpython/issues/53646