[SYCL][NativeCPU] Build libclc target-independently. (#17408)
This change creates a Clang NativeCPU target to be used for creating
LLVM IR for NativeCPU, builds libclc with this target, and does the
initial processing of NativeCPU device code with this target.

This allows the libclc to be independent of the concrete host target,
allowing libclc to be re-used across different host targets, and across
different ABIs for the same host target.