[UR] Fix OpenCL handles generated from sub-devices
Sub-devices don't appear in the list of devices, and so the prior OpenCL
implementation would assume that they are invalid. With this patch,
creating a sub-device will add it to a "cache" stored in the platform.
Multiple `DeviceCreateWithNativeHandle` calls with the same native
handle will result in the same UR handle. A test has been added to
verify this.

This unfortunately required shuffling headers around to avoid circular
imports.

This fixes an issue tracked internally by Intel as URT-903