[SYCL][Docs] Add SYCLBIN feature and format design document (#16872)
This commit adds a design document detailing the SYCLBIN binary format
for representing SYCL device kernel binaries to be loaded dynamically at
runtime. Additionally, the design document details how this is to be
handled by the SYCL runtime, driver and clang tooling.

As the design of SYCLBIN files relies heavily on the property sets, this
PR also adds documentation to the existing property set functionality.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>
Co-authored-by: Michael Toguchi <michael.d.toguchi@intel.com>
Co-authored-by: Greg Lueck <gregory.m.lueck@intel.com>
Co-authored-by: Alexey Sachkov <alexey.sachkov@intel.com>