Do not vectorize NoInline kernels.
A recent LLVM 21 change prevents us from being able to vectorize
NoInline kernels, but it was probably never a good idea to vectorize
them anyway: it is not clear whether the vectorization we perform is
something that a user who writes `__attribute__((noinline))` would want.
Therefore, explicitly reject vectorization of NoInline functions and
restore affected tests to their original version that permitted them to
not be vectorized.