[SYCL] reset enqueue status so that buffer can be used after failure is addressed (#17871)
CMPLRLLVM-65738 - If we run out of memory, an exception is thrown and
can be caught by the user application. They can address this in several
ways: free memory, .wait() on outstanding operations, etc. But despite
addressing the problem, SYCL has a bug wherein the buffer is not usable.
This is because the EnqueueStatus isn't reset, the Command/Requirement
cannot be re-enqueued.
The fix here is straightforward. We also improve the FailedDependency
unit test.