[UR] Add UR_DEVICE_INFO_VALIDATES_ON_LAUNCH
This device info is used to specify whether a given device can validate
arguments passed to a kernel when it is launched. This also updates the
spec to specify that if this value is not set, then running a kernel
with invalid arguments is UB.

This does not imply any behavior changes in adapters themselves; they
already UB-ed anyway. This just writes it into the spec and makes
some conformance tests not fail them for it.