[UR] Move non-conformance/adapter tests to lit (#17580)
This adds support for lit to unified-runtime, in the form
of a new target `check-unified-runtime`. This target is available
when UR is built standalone and when built as part of dpcpp.

In addition, there are targets like `check-unified-runtime-unit`
which run a subset of the lit tests.

When built standalone, `lit` and `filecheck` must be avilable in
the path. A `requirements_testing.txt` file is available to
install test dependencies.

When built as part of dpcpp, a new cmake variable is available;
`SYCL_UR_BUILD_TESTS` which is off by default. When enabled,
`check-unified-runtime` will be available as a target and will
also run as part of `check-all`.

All tests have been ported with the exception of:
* Adapter tests
* Conformance tests

The Python version for macos was also bumped to 3.10 to enable
use of a newer filecheck version.