[SYCL] Stop emission of modf intrinsic for NVPTX/AMDGCN
A recent change in #126750 changed the lowering of the modf builtin(s)
to lower to the llvm.modf intrinsic. Certain SYCL targets such as NVPTX
and AMDGCN cannot currently handle this intrinsic in the backend because
they don't have the necessary target library info. Even if they did, the
SYCL device libraries are linked in earlier at the IR level so we'd be
left with an unresolved symbol.

We have been working around this at clang codegen time by avoiding the
emission of intrinsics altogether by (ab)using a guard on math
intrinsics.

The problem with us hooking into GenerateIntrinsics like this is that
SYCL is wanting to use it as a guarantee that (a certain subset of)
intrinsics won't be emitted if GenerateIntrinsics is false, whereas
upstream does not make this guarantee: it's an opt-in toggle for a more
optimal lowering strategy. Thus we're always going to be liable to this
sort of upstream change.

This doesn't feel like the right mechanism for SYCL to handle these
builtins (or intrinsics) long term, but this fix restores the previous
behaviour without making things much worse.

Fixes #17813