[SYCL][Test] Replace multi-device kernel_bundle e2e tests with unit test (#17691)
Replace the following e2e tests with unit test:
```
    ProgramManager/multi_device_bundle/device_libs_and_caching.cpp
    ProgramManager/multi_device_bundle/build_twice.cpp
    KernelAndProgram/persistent-cache-multi-device.cpp
```
    
The reason is that the checks performed in those e2e tests (number of UR
calls, etc.) can be done more effificently using unit tests, such checks
in e2e tests are expensive. Another reason is that e2e tests use
NEO-specific `CreateMultipleRootDevices=[num]` environment variable
which is not guaranteed to work, in unit tests we can emulate as many
devices as we want.