[SYCL] Use urGetDeviceInfo to check bfloat16 native conversion extension (#18023)
As agreement in https://github.com/intel/llvm/pull/16729 , In general we
discourage the use of the extension strings as it is a remnant of the
OpenCL days. Instead we want to use urDeviceGetInfo when possible. This
is a follow-up pr to replace checking OpenCL extension string with
urGetDeviceInfo for native bfloat16 conversion extension.

---------

Signed-off-by: jinge90 <ge.jin@intel.com>
Co-authored-by: Steffen Larsen <steffen.larsen@intel.com>