[UR] For non-standalone builds, use built DPCXX/Extractor
If we are not a standalone build (that is, we are built as part of
intel/llvm) and `UR_DPCXX` is not specified as part of the cmake
options, we use the `clang` built by `intel/llvm` to generate device
binaries.

Likewise, if no `UR_DEVICE_CODE_EXTRACTOR` is specified and we are
not standalone, the `clang-offload-extractor` from intel/llvm is used.

This change should not affect CI, but does mean that the UR testing
targets can be built in an intel/llvm local build.