[SYCL][UR][L0 v2] fix use after free on queue
It is possible for queue to be released before all events
associated with that queue are released. To avoid use-after
free, avoid calling any function on queue and instead rely
on a device handle (which will be alive until adapter is destroyed).

The only place where we still call a function on the queue is
in event release in case the timestamp is not ready.
This should be safe - if the queue is released then the timestamp
must have been written to the event already so this path should
never be taken. However, there is a potential race: if
getEventEndTimestamp returns false and just after that
the queue finishes and is dstroyed. To avoid that, retain
the queue when starting the recording, and release whenever
profiling data is destroyed (on resetQueue and event release).