[CI][E2E] Run Windows Battlemage with prebuilt binaries (#18042)
Also disable tests that were flaky hanging in this configuration.

When running fully on BMG tests took around 15min. With this change
building e2e tests takes around 10m, and running the tests took around
6m. On paper this isn't any faster, however since we are spending less
time on the BMG machine (which we currently only have one of in CI) this
should improve the queue times for it. Additionally there is further
opportunity to speed up the e2e build stage by moving it into the same
job that builds the compiler, like we do for Linux.