[libspirv][AMD] Fix sub-group shuffle builtins
Commit 316418d modified the AMDGCN sub-group shuffle builtins. In doing
so it broke them in a couple of ways.

The first is a typo that hid the SubgroupShuffleXorINTEL builtins for
sub 32-bit types but ommitting the 'INTEL' part of the builtin name.

The second was an issue of mangling for the 'half' data type, which
accidentally stopped the SYCL host being able to find the right symbols.

Since libspirv is defined in OpenCL C, the plain 'half' data type is
mangled differently to the sycl::half type used by SYCL: 'Dh' vs
'DF16_'. We've long worked around this using manual mangling of the
builtins, which has never been very nice and has often caused bugs that
are awkward to track down.

This PR takes an alternative approach and instead uses a custom typedef
to the _Float16 type, and defines the sub-group shuffle builtins using
that new type. Since _Float16 is the 'half' type used everywhere except
OpenCL, we arrive at the correct mangling. If this approach proves
successful we might be able to expand on it and remove much of the
manual mangling elsewhere in libspirv.