[SYCL] Cache device aspects during `shared_ptr<device_impl>` creation
Computing some aspects requires string comparison or even heap
allocations of strings (for `device_impl::has_extension`) and caching
that on lower levels (e.g., UR) doesn't look reasonable. On the other
hand, SYCL aspects don't change for the entire lifetime of a device, so
it makes sense to compute them just once.

The overhead of such compute is about 36us on my machine. It's
significant compared to the `device_impl` creation, but we do create all
the devices hierarchy together with the platforms and those are
orders of magnitude more expensive than computing the aspects (~100ms).

Implementation-wise there are two things worth mentioning:

  * `device_info.hpp` APIs take `DeviceImplPtr` and not `device_impl`
    ptr/ref and are needed to compute some of the aspects. As such, we
    can't pre-compute everything in the `device_impl` ctor and have to
    do that in `platform_impl::getOrMakeDeviceImpl`
  * Mock devices in unittests don't implement queries for all the
    aspects, so pre-computing them leads to crashes. I've decided to
    simply disable such caching for the devices named "Mock device".