[SYCL][UR] fix use after free on queue in eventRelease
There was a potential race: if getEventEndTimestamp returns
false and just after that the queue finishes and is destroyed,
we would call deferEventRelease on invalid queue.

This patch reworks the logic for deferring event release.
Instead of registering the events to be released in the queue,
we just clear adjustedEventStartTimestamp and adjustedEventEndTimestamp
to mark the event as not timestamped. Since events are never
actually destroyed, only returned to the pool, any pending
timestamp write will always succeed. Since we are only dealing
with in-order queues, the timestamp can only increase.