return UR_RESULT_ERROR_UNSUPPORTED_ENUMERATION in more cases
Because CL_INVALID_VALUE is returned both when the size of the value
being queried is incorrect and when the information being queried is
unrecognized, it is difficult to differentiate between the UR
INVALID_SIZE and UNSUPPORTED_ENUMERATION errors.  For now, we will
check for and return INVALID_SIZE first, which is what the tests are
expecting.