[SYCL] Disable kernel stub for now (#28629)
642481a4286c9006958274531ee173b347866c50 start to emit kernel stub for
opencl, it affect the SPIR calling converision.

However, the upstream only emit stub in GenerateCode, sycl code is
emitted in EmitDeferred instead, so we end up not generating stub and
only changed the kernel calling conversion.

This PR disable the change to kernel calling conversion for now.
If we want to enable kernel stub for sycl, we should follow up to emit
the stub for sycl as well.