[Driver][NFC][FPGA] Remove remaining FPGA offload support in driver (#18087)
Following
https://github.com/intel/llvm/commit/491915b2ac92caf76f611e804644abce30e148ae,
remove the remaining items in the driver that support FPGA.

Small exceptions to this include any specific diagnostic behaviors that
inform the user of option usage that is not supported due to the removal
of FPGA. Cleanup of those items can occur following a later major
release. The extension enabling options for the `sycl-post-link` tool
specific for fpga also remain for now.