[SYCL][BINDLESS] Fix mem_image_free for gather array (#18109)
The case of `image_type::gather` would lead to an exception being thrown
in `free_image_mem`/ `std::terminate` in `~image_mem_impl`. I missed
this in https://github.com/intel/llvm/pull/17322.
This fixes that.

Note that this doesn't fully address
https://github.com/intel/llvm/issues/16331 since the call to UR could
still throw in theory. However, in order to avoid such a UR call in a
destructor the implementation would have to be completely changed. As it
is, if for some unusual circumstance the image free did fail in UR, I
think there is at least an argument it be better that this does lead to
`std::terminate`. So I'm leaving this for discussion.

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>