[SYCL] Use defined(__SPIRV__) instead of !defined(all other targets) (NFC)
This is a simpler logic, the `#else` branch already assumes built-in SPIR-V
variables anyway, so any future target would not want to take it.
Also this removes one point that needs to be modified if new targets were
to be added.

NOTE:
I could have flipped the logic to make the `#if` positive, but I didn't
to make this a smaller diff, let me know if that would be preferable.