[UR][OpenCL][Graph] Set simultaneous-use property
The merged PR https://github.com/intel/llvm/pull/17658 added the
requirement for OpenCL cl_khr_command_buffer implementations to support
the simultaneous use capability in order to support UR command-buffers.

There was an oversight in this PR where the simultaneous use property
wasn't set on command-buffer creation in the OpenCL adapter. This is
rectified in this patch.