[UR][CUDA] Remove vtable from ur_exp_command_buffer_handle_t_ (#18138)
With the handle removal, we will start requiring handle types to not
have
vtables. This changes the `ur_exp_command_buffer_handle_t_` type so that
the type is encoded as an enum rather than a specific type.

Three of the specific handle types require extra data, which is now
stored in another field as a std::variant.