[libspirv] Use the remangler to mangle half types (#18083)
This commit revisits the way we achieve the desired host-side mangling
of `half` types as `DF16_` (from the `_Float16` type) coming from the
original mangling of `Dh` type that OpenCL produces. We were previously
manually achieving this by writing over a thousand wrapper functions
from `_Float16` types to `half` types.

The remangler can just as easily do this for us, while reducing the
total source code line count by over 8000.

This work was originally spurred on by commit 316418d, which modified
the AMDGCN sub-group shuffle builtins. In doing so it broke them in a
couple of ways.

The first is a typo that hid the SubgroupShuffleXorINTEL builtins for
sub 32-bit types but ommitting the 'INTEL' part of the builtin name.
That was easily fixed by adjusting the builtin name.

The second was an issue of providing builtins using the OpenCL 'half'
data type, without the `_Float16` -> `half` wrappers which we relied on.
This accidentally stopped the SYCL host being able to find the right
symbols.