[SYCL][Coverity] Fix MISSING_LOCK coverity hits in the program manager (#18121)
This PR guards accesses to `m_DeviceImage` and `m_VFSet2BinImage`
caches/maps in the Program Manager using `m_KernelIDsMutex` in
situations where multiple threads may be accessing the maps at the same
time.

I am opting to use `m_KernelIDsMutex` here as both maps are commonly
accessed while holding said mutex: In fact, a lot of maps/caches in the
program manager use `m_KernelIDsMutex` (This is not the most optimal
solution, but that is out of the scope of this PR, see #18165)

This PR addresses Coverity hits that needs to have fixes submitted
before the code cutoff.

---------

Co-authored-by: Steffen Larsen <steffen.larsen@intel.com>