[SYCL][Graph][OpenCL] Map copy/fill to SVM
There is currently no cl_khr_command_buffer or cl_intel_unified_shared_memory
entry-point for appending USM copy or fill commands to a command-buffer.
This prevents these commands from being added to a graph for the OpenCL
backend.

The long term solution to this is the OpenCL USVM extension which will
align USM and SVM, allowing the existing SVM entry-points to be used.
To prepare for this, map the UR entry-points to the
cl_khr_command_buffer SVM copy/fill commands.

This will work on OpenCL implementations that share an implementation
for USM and SVM.

Signed-off-by: Ewan Crawford <ewan@codeplay.com>