[UR][Graph] command-buffer unsupported OpenCL CTS tests
There are currently limitations on what commands can be added to
a SYCL-Graph that we document in the [design doc](https://github.com/intel/llvm/blob/sycl/sycl/doc/design/CommandGraph.md#opencl)
and mark as unsupported in Graph E2E tests.

However, we never marked the equivalent UR command-buffer commands as
unsupported in the CTS, so they will currently fail when testing against
an OpenCL backend with support for the extension. Address this by
marking CTS tests which use these commands as unsupported.