[Driver][SYCL] Remove object upon failure
Typical behaviors when the compilation fails is to have the resulting
output object to be removed. Due to the fact that there are multiple
compilations that occur during an offload compilation, the actual final
output object may not be directly associated with the compile causing it
to not be removed.

Update the cleanup file list to remove the output result file regardless
of the associated job action when we are performing offload.

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>