[SYCL] Hide inline definitions of stdio functions (#18174)
MSVC's inline definitions of stdio functions such as printf() are only
meant for host code; device code cannot call them. Device code can,
however, call ext::oneapi::experimental::printf() which depending on the
target is implemented as a call to the global printf(). The availability
of a definition of the global host printf() makes it more difficult to
handle in device code, so hide it.