[UR][L0] Track when offsets are used by kernels
The vast majority of kernels are expected to never have an offset,
because the offset feature was deprecated in SYCL 2020.

A small number of kernels will still have an offset. Since this case is
uncommon, it is less important to optimize.

There is a pathological case where the same kernel alternates between
different overloads of parallel_for with and without an offset. Keeping
track of whether the last submission had an offset is intended to
address this case, while still allowing us to skip the L0 call.

Signed-off-by: John Pennycook <john.pennycook@intel.com>