[SYCL] Add fallback support for KernelCompiler loading of ocloc (#18245)
Presently, the KernelCompiler just tries to load whatever ocloc shared
library is in the environment. And if it doesn't find a compatible one,
then it reports that OpenCL C compilation is not supported. But on
Windows this should be more robust, so we are adding fallback support.
If the ocloc shared library in the environment isn't compatible, try the
one in the OneAPI installation.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>