[UR][CUDA][HIP] Fix error handling in bindless images code (#18246)
`UR_CHECK_ERROR` throws an exception, so it must only be used in
`try/catch` blocks.

This patch fixes the uses of `UR_CHECK_ERROR` in the image code with a
mix of adding missing `try/catch` blocks and replacing it with
`UR_CALL`.

Also skips checks in a catch block trying to cleanup resources before
returning an error, we shouldn't re-throw from that and returning the
initial error is the more helpful thing to do.