[Driver][SYCL] Do not set workaround predefines with /MDd (#18252)
We are currently setting 2 pre-defines which workaound a compilation
issue: -D_CONTAINER_DEBUG_LEVEL=0 and -D_ITERATOR_DEBUG_LEVEL=0. We can
no longer rely on these macros as _CONTAINER_DEBUG_LEVEL will be going
away with a future release of Visual Studio.

The issues that this worked around are detailed in our release notes as
a known issue:
https://github.com/intel/llvm/blob/aa2c876f529f942943afaf9395694fe07a7f6121/sycl/ReleaseNotes.md?plain=1#L2700

Remove these workarounds in preparation for the upcoming MSVS release.