[UR][Graph] Serialize submissions of the same command-buffer
Update the UR command-buffer enqueue semantics such that
command-buffer submissions are ordered with a dependency on
previous submissions of the same command-buffer.

The is the same semantic as we currently specify for executable
graph objects in the SYCL-Graph extension. Allowing for a more
efficient SYCL-RT implementation on top of UR where less UR events
need to be created (future work).