[SYCL] Simplify secondary queue usage
The requirement of SYCL submit functions taking a secondary queue does
not mandate more than that the runtime must attempt to enqueue onto the
secondary queue if the it fails to enqueue on the primary queue. Using
this, the SYCL implementation can be simplified to simply retry the
enqueue if the enqueue to the primary queue throws an exception.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>