[LIBCLC] Drop AMD code object version from libclc (#18294)
Upstream recently bumped the default AMD code object version to 6, which
is only supported by ROCm 6.3 and above.

There is a flag `-mcode-object-version=` to make the compiler use an
older version. But before this patch it broke at the libclc linking
stage since libclc had the default value in a module flag.

So this patch makes the prepare builtin tool remove this module flag
from the libclc bitcode libraries.