[L0][UR] Do not set global offset unless required (#18242)
Previously, zeKernelSetGlobalOffsetExp was called for every kernel
launch. The vast majority of kernels are expected to never have an
offset, because the offset feature was deprecated in SYCL 2020, and we
should optimize for this case.

The SYCL RT currently passes {0, 0, 0} in the case where there is no
offset. To optimize this case:
- A zero offset is treated equivalently to a NULL offset, and
zeKernelSetGlobalOffsetExp is not called.
- A non-zero offset triggers a call to zeKernelSetGlobalOffsetExp before
launching the kernel.
- A non-zero offset triggers a call to zeKernelSetGlobalOffsetExp after
launching the kernel, to reset the offset to zero.

This will introduce additional overhead to the uncommon case where
offsets are specified, but we plan to remove this anyway.

In the long-term, the check for a {0, 0, 0} offset should probably be
moved into the SYCL headers and NULL should be passed directly to UR.
However, this will require wide-reaching changes to other UR adapters
and the UR specification.

---------

Signed-off-by: John Pennycook <john.pennycook@intel.com>