[SYCL] Use raw pointers in enqueue flow (#18306)
For `kernel_impl` parameter - it's an optional input only parameter that
is owned by the caller for the duration of the call and we don't extend
its lifetime/create copies as can be seen by the absence of changes in
the callee (all we need is to access its data).

For `event_impl` "out" parameter the semantics is such that the caller
creates an object, and the modified interfaces update that object with
the UR handles once enqueue/launch is done. The ownership remains
controlled by the caller.

---------

Signed-off-by: Sergei Vinogradov <sergey.vinogradov@intel.com>