[CI] Rename compiler variables (#18338)
We have `cxx` and `cxx_compiler`, very similar names so it can be
confusing. `cxx` is used to build the project, while `cxx_compiler` is
the compiler produced so renaming the latter to `sycl_compiler`.
Additionally aligning the Windows with the Linux build job by renaming
`compiler` to `cxx`.

Addresses comments from #18322