[SYCL][E2E] Fix LIT requires vulkan directive
The addition of "vulkan" to `config.available_features` was moved to
`bindless_images/lit.local.cfg` in #18220. This broke the
`// REQUIRES: vulkan` directive, and tests were reporting as being
unsupported despite vulkan being available on the system.

This patch moves the check and addition of vulkan to the available
features back to `test-e2e/lit.cfg.py` to fix the issue.