[SYCL] Eliminate XPTI overhead when it is disabled (#18334)
Even if the XPTI collector is not used, we still have some overhead
related to XPTI instrumentation when we call `xptiRegisterStream`
function.

With these changes, we cache the result of the `xptiTraceEnabled`
function in the static variables. It is safe to cache it because XPTI
collector is enabled by the `XPTI_FRAMEWORK_DISPATCHER` environment
variable before the process starts.

VTune shows ~2% decrease in the number of instructions retired on a
submit kernel path.

---------

Signed-off-by: Sergei Vinogradov <sergey.vinogradov@intel.com>