[SYCL] Add hasSpecialCaptures() constexpr function
Detecting which kernel functions have special captures (i.e., anything that
isn't a standard layout class or pointer) will enable us to introduce a fast
path for those kernels.

Note that supporting this functionality requires all of the kernel descriptor
functions to be constexpr. This was already the case for the functions
generated for the integration header, but was not true for some of the new
builtins and placeholder functions in kernel_desc.hpp.

Signed-off-by: John Pennycook <john.pennycook@intel.com>