[SYCL] Extract args directly from kernel if we can
In some cases, all values that need to be passed as kernel arguments are stored
within the kernel function object, and their offsets can be calculated using
the integration header or equivalent built-ins. In such cases, we can therefore
set kernel arguments directly without staging via MArgs.

This first attempt is limited to the simplest cases where all kernel arguments
are either standard layout types or pointers. It may be possible to extend this
approach to cover other cases, but only if some classes are redesigned.

The implementation currently stores some information (e.g., the number of
kernel arguments) inside of the handler, because there is no way to access
the kernel type within handler::finalize().

Signed-off-by: John Pennycook <john.pennycook@intel.com>