[CI] Test on all devices when not specified  (#18356)
Nightly https://github.com/intel/llvm/actions/runs/14892412297

We forgot to set `target_devices` when moving the nightly Windows
run-tests job over to the composite workflow, but we can't require it to
be set because it needs to be unset for `build-only` mode, so just use
`all` if not specified and fix the nightly.

Currently we get a CMake error because `sycl_devices` is empty which
fails to be parsed.

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>