[SYCL][E2E] Enable llvm-lit to accept device architecture (#18197)
Allow for selection of specific devices via
`level_zero:arch-intel_gpu_bmg_g21` which accepts device architectures
listed under `sycl-ls --verbose`. Devices with multiple GPU's under the
devices with multiple GPU's under the device selection style of
`backend:*device type*` such as `level_zero:gpu` can cause incorrect
aspects to be marked as unsupported in the tests ran through `llvm-lit`
using `// REQUIRES: aspect-`. In addition, only the first device listed
via `sycl-ls` under the label is ran.
    
This occurs because using the device selection style of `level_zero:gpu`
there is currently no way to choose a specific device that falls under
that category. Such as a battlemage GPU and an intel iGPU.
`sycl/test-e2e/lit.cfg.py` takes all available device aspects and marks
aspects as supported that are available for all the devices. This is
problematic for aspects such as `ext_oneapi_bindless_images` which is
supported on battlemage but not Intel iGPU's. Even if battlemage, which
supports bindless images is used to run all the tests as it comes first
in the device list shown under `level_zero:gpu`, tests requiring that
aspect will not run.