[SYCL] Add hasSpecialCaptures() constexpr function (#18386)
Detecting which kernel functions have special captures (i.e., anything
that isn't a standard layout class or pointer) will enable us to
introduce a fast path for those kernels.

Note that supporting this functionality requires all of the kernel
descriptor functions to be constexpr. This was already the case for the
functions generated for the integration header, but was not true for
some of the new builtins and placeholder functions in kernel_desc.hpp.

---

As a note to reviewers: I do have some prototype functionality ready
that uses this, but I thought that splitting things into two separate
pull requests would make things easier to review.

---------

Signed-off-by: John Pennycook <john.pennycook@intel.com>