Defer arg extraction until handler::finalize
Rather than extracting arguments from the lambda when the kernel is enqueued,
store a pointer to the lambda alongside relevant information from the
integration header or compiler builtins.

Storing this information will allow us to defer the extraction of arguments
until we reach handler::finalize(), at which point it may be possible to
set the kernel arguments directly without populating MArgs.

Signed-off-by: John Pennycook <john.pennycook@intel.com>