[SYCL] Deprecate discard_events queue prop, make impl no-op (#18059)
Deprecates sycl_ext_oneapi_discard_queue_events in favour of the
sycl_ext_oneapi_enqueue_functions extension.
In order to simplify/allow further simplifications to the
handler/scheduler code the implementation of the `discard_events` queue
property is also made a no-op immediately.

The main consequence of this is that users of `discard_events` will have
to ensure they use the new `submit_without_event` / `nd_launch` etc APIs
from
https://github.com/intel/llvm/blob/sycl/sycl/doc/extensions/experimental/sycl_ext_oneapi_enqueue_functions.asciidoc
`sycl::event` is very lightweight already in the level_zero v2 adapter
and discard_events already doesn't have a very big performance advantage
for this case. Therefore the biggest consequence of this PR from a
performance point of view is in the cuda/hip backends. Users of these
backends need to be aware of the performance implications of making the
`discard_events` no-op, and should ensure that they switch to using the
`sycl_ext_oneapi_enqueue_functions`.

This PR leaves all other functionality working correctly.

---------

Signed-off-by: JackAKirk <jack.kirk@codeplay.com>
Co-authored-by: Peter Žužek <peter@codeplay.com>