[UR][CUDA] Cleanup UMF pools creation code (#18361)
This code was roughly duplicated between the context and the platform,
and could also be cleaned up and refactored to be clearer and not
require as many helper functions and macros.

It is going to create more "MemoryProviderParams" during platform
initialization. But that cost of that is likely pretty small and not on
the hot path, so it seems like a fair tradeoff for cleaner code.