Use native for the system separator for source path string (#3157)
Previously it used posix, causing strange behaviour like reported in
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/issues/3126

It's NFC change for the translator-based tool chains. Yet if SPIR-V
consumer is not llvm-spirv - the fix should help.

Note, the patch is not attempting to guess separators depending on the
user's IR input, so for the linux <-> windows cross compilation the
result will be odd (like it was before).

Signed-off-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/c59f0f48ac98325