[UR][Test] Explicit target name in KernelsEnvironment::LoadSource (NFCI) (#18279)
Allow explicitly overriding the name of the target to load kernel
sources for. This is useful if the backend supports multiple targets,
and a test wants to load the source for a specific one. Multiple targets
might for example be SPIR-V and native binary format for the OpenCL or
Level Zero backend, or future vendor specific intermediate
representations.

The overload taking a platform is unchanged, it will load a the source
for the "default" target for the backend, which is the same as before.