[NFCI][SYCL] Use `get_info_impl` (not `_nocheck`) for USM info descriptors
I tracked the appearance of `_nocheck` down to
https://github.com/intel/llvm/pull/937 and James doesn't remember
anything specific requiring that vs "checked" version. Also, some
similar USM-related aspects (those that don't just delegate to
`get_info`) do use "checked" APIs, so it seems logical to unify this
processing.

The main reason I do this is that it would be a bit easier to cache the
values of "checked" interfaces by pre-computing USM support in
`device_impl`'s ctor. Note that we perform at least some of those every
time we allocate USM memory, so doing the caching is desirable (even for
the sake of cleaning up our traces).