[SYCL][CL] Fix ownership of native handle in sycl::make_queue (#18309)
Level Zero is the only backend which has any explicit ownership transfer
controls when creating SYCL objects from native handles.

The CUDA and HIP backends already had the default to not transferring
ownership of native handles.

Up until this patch the OpenCL backend was transferring ownership when
creating with a native handle, when combined with the changes to the
OpenCL adapter in https://github.com/intel/llvm/pull/17572 this caused
crashes in existing workloads which expect to manage ownership of the
OpenCL native handles.

Fixes URT-905.