Revert "[SYCL] Refresh hardening flags applied to the project " (#18472)
This reverts commit dda1c1598a69e311c36938466528031f4025541e.

Validation of this commit fell to a human factor and as such haven't been properly performed. From what we see, existing flags are enough to pass various necessary binary scans, so the flags will be re-instated later after we carefully evaluate failures caused by them - for that we will need to start using them on the sycl branch first to ensure that they are regularly tested there as well.