[SYCL][E2E] Use `ROCM_PATH` from the system environment (#18468)
For ROCm installations that aren't just in `/opt/rocm`, `clang` uses
`ROCM_PATH` at runtime to find the ROCm device libraries, and this is
done even in regular SYCL compilation. So this is needed for lit tests
in certain configurations.

This was accidentally removed in
https://github.com/intel/llvm/pull/17692 it doesn't cause any issues in
the CI because the CI has ROCm installed in the standard `/opt/rocm`,
but it causes issues on local setups.

Also remove `--rocm-path` on Windows, this should be covered by
`ROCM_PATH`.