[Bindless][SYCL][UR] Create a sampled image with a single UR API (#18384)
We no longer need to create a UR sampler object in order to create a
sampled image. The backends can simply use the sampler description to
create the sampled image.

Note this is an ABI break in the currently still experimental bindless
images extension for Unified Runtime.

Addresses the following issue:
https://github.com/oneapi-src/unified-runtime/issues/1463
Relevant discussion from
https://github.com/oneapi-src/unified-runtime/pull/2640#discussion_r1942467162

Additionally, avoiding the allocation of a sampler object itself fixes
memory leak(s) in all adapters where the UR sampler handle was never
released which was caused by the use of the API in the SYCL RT only
calling urSamplerCreate and not Release for the sampler associated with
the image. Also, while HIP and Cuda do not further create backend
specific sampler objects in the respective adapters, Level Zero was
creating one that never got released upon freeing the image memory which
lead to leaking the Level Zero sampler memory as well. Now that Level
Zero can create a sampled image from a descriptor only, the sampler
object is not needed altogether, allowing to rid of the above mentioned
leaking memory allocations and simplify the runtime.