[Driver][SYCL] Pass as system headers for external SYCL host compilation (#18449)
The use of `-fsycl-host-compiler=arg` allows a user to specify a 3rd
party compiler to perform host compilation when using `-fsycl`. The
compilation will also include the SYCL specific headers. Pass in these
headers as system headers with `-isystem` so they are treated as system
headers as opposed to user headers.

---------

Signed-off-by: Michael D Toguchi <michael.d.toguchi@intel.com>