[XPTI][INFRA] Fixes minor standalone build and code issues (#18452)
This PR is first of many XPTI changes to address the performance issues
that have been reported.

(1) xpti::Framework::instance() - testing the use of std::call_once
against current implementation
(2) Added cleanup for default stream "xpti.framework"
(3) Switched to EmHash instead of std::unordered_map that improve
notification performance by 10%

Most of the performance overheads are from xptiRegisterObject() and
xptiAddMetadata(). This requires changes in the SYCL runtime an will be
addressed in a subsequent PR

---------

Signed-off-by: Vasanth Tovinkere <vasanth.tovinkere@intel.com>