[SYCL] Revert prepare-builtins change for non-NativeCPU. (#18490)
In #17999 I changed prepare-builtins to clear target-cpu and
target-features for NativeCPU, and attempted to preserve existing
behavior for other targets. For NativeCPU, two errors cancelled each
other out so we do the right thing there, but for other targets, that is
not the case and target-cpu and target-features were also being cleared
unintentionally.