[NFC] Unsupport 2 HLSL test on windows
Both are failing
https://github.com/intel/llvm/actions/runs/15055364179/job/42319925341?pr=18403

FAIL: Clang :: Driver/dxc_dxv_path.hlsl (11426 of 22048)
 note: command had no output on stdout or stderr
 error: command failed with exit status: 0xc0000135

The error "command failed with exit status: 0xc0000135" on Windows indicates
that an application or command could not start because a required system component-typically
the .NET Framework-is missing, corrupted, or not enabled.
This is a common issue when running applications built with .NET,
especially if the necessary version of the framework is not installed or enabled on your system.

These two tests were passing on our internal windows testings,
so looks like env issue. Unsupported them for now.