[SYCL][COMPAT]  Fix error on headers, add helloworld test (#18401)
This PR fixes an error on the SYCL compat headers that caused build to
fail.
The error manifests when building the basic hello world from the
documentation.
To prevent the basic hello world from failing again, this PR adds an e2e
for the basic hello world
and updates the documentation to match the updated hello world source
code.